pyproject.toml
[tool.poetry]
name = "qwen2-vl-api"
version = "0.1.0"
description = ""
authors = ["maguowei <imaguowei@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
qwen-vl-utils = "^0.0.4"
modelscope = "^1.17.1"
transformers = {git = "https://github.com/huggingface/transformers"}
accelerate = "^0.34.0"
torchvision = "^0.19.0"
ipython = "^8.27.0"
fastapi = {extras = ["standard"], version = "^0.112.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
